study04/requirements.txt
# Created: 2026-03-24 22:12:00
flask
requests
python-dotenv
PyJWT
bcrypt
